+++ Matrix/Matrix/utils/matrix/Matrix.hpp
@@ -4,15 +4,15 @@
 
 class Matrix {
  public:
-  Matrix(std::string matrix_name = "Matrix", int dimension_x = 1,
-         int dimension_y = 1);
-  Matrix(std::string matrix_name, int dimension_x, int dimension_y,
+  Matrix(std::string matrix_name = "Matrix", size_t dimension_x = 1,
+         size_t dimension_y = 1);
+  Matrix(std::string matrix_name, size_t dimension_x, size_t dimension_y,
          std::vector<std::vector<float>> arr);
   ~Matrix();
 
   Matrix(const Matrix& src);
-  inline int get_x_dimension() const { return dim_x; }
-  inline int get_y_dimension() const { return dim_y; }
+  inline size_t get_x_dimension() const { return dim_x; }
+  inline size_t get_y_dimension() const { return dim_y; }
   inline std::string get_matrix_name() const { return matrix_name; }
   inline void set_matrix_name(std::string str) { matrix_name = str; }
 
@@ -23,7 +23,7 @@ class Matrix {
 
  private:
   float* matrix{nullptr};
-  int dim_x, dim_y;
+  size_t dim_x, dim_y;
 
   std::string matrix_name;
